refactor: 컴퓨터 랜덤값 생성 함수 ComputerInput 파일로 분리
